feat: 상품상세조회 api 기본 응답 구현#33

+++ src/main/java/spharos/msg/domain/product/entity/Product.java
@@ -33,7 +33,7 @@ public class Product extends BaseEntity {
 
     @NotNull
     @Column(columnDefinition = "integer default 0")
-    private Integer defaultImageIndex;
+    private Short defaultImageIndex;
 
     @NotNull
     @Column(columnDefinition = "integer default 0")
